requirements.txt
ollama>=0.3.0
sentence-transformers>=3.0.0
anthropic>=0.40.0
pydantic>=2.0.0
pytest>=8.0.0
pytest-timeout>=2.3.0
requests>=2.31.0
numpy>=1.24.0
fastapi>=0.110.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
